+++ src/test/converter2/issues/gh2999.ts
@@ -1,0 +1,10 @@
+/**
+ * A test object with property a.
+ */
+const LocalObject = {
+    a: 1,
+};
+
+export const Options = {
+    LocalObject,
+};